+++ packages/widget/src/components/FlexJarGuidePanel/index.ts
@@ -1,0 +1,4 @@
+export {
+  FlexJarGuidePanel,
+  type FlexJarGuidePanelProps,
+} from "./FlexJarGuidePanel.js";