+++ packages/widget/src/index.ts
@@ -1,4 +1,8 @@
 export * from "./modal/index.js";
 export * from "./components/questions/index.js";
 export * from "./core/index.js";
+export {
+	FlexJarGuidePanel,
+	type FlexJarGuidePanelProps,
+} from "./components/FlexJarGuidePanel/index.js";
 export type { FlexJarRenderQuestionProps } from "./types.js";