+++ packages/widget/src/modal/index.ts
@@ -1,6 +1,7 @@
 export {
   FlexJarModal,
   type FlexJarModalProps,
+  type FlexJarModalWidth,
   type FlexJarFollowUpQuestion,
   type FlexJarMainQuestion,
   type FlexJarSurveyConfig,